rename css file & rearrange css file(assign to component file)
